requirements.txt
absl-py==1.1.0
albumentations==1.1.0
cachetools==5.2.0
certifi==2022.5.18.1
chardet==4.0.0
cloudpickle==2.1.0
dataclasses==0.6
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
gym==0.24.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
pandas==1.4.3
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.26
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
six==1.16.0
sklearn==0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.9.1
tensorflow-model-optimization==0.7.5
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
nvidia-tensorrt==7.2.2.1 # for TF TensorRT inference
tensorrt==1.3.0 # with --no-deps for PyTorch TensorRT
torch-tensorrt # with --no-deps for PyTorch TensorRT